--- quinzena6/LabenuForms/labenuforms/src/components/inputsDeTexto.jsx
@@ -1,19 +1,0 @@
-import React from "react"
-
-export default class InputsDeTexto extends React.Component {
-   
-
-    render () {
-
-        return (
-            <div>
-                <div>
-                <p>{this.props.texto}</p>
-                <input type="text" >
-                 </input>
-                </div>
-            </div>
-        )
-
-    }
-}
