+++ quinzena6/LabenuForms/labenuforms/src/components/pagina2.jsx
@@ -1,0 +1,30 @@
+import React from "react"
+
+export default class SelectItems extends React.Component {
+
+    render () {
+
+        return (
+            <div>
+                <h1>Etapa 2 - INFORMAÇÕES DE ENSINO SUPERIOR</h1>
+                <div>
+                    <p>Qual Curso?</p>
+                    <input type="text" >
+                    </input>
+                </div>
+                <div>
+                    <p>Qual a unidade de ensino?</p>
+                    <input type="text" >
+                    </input>
+                </div>
+            </div>
+        )
+
+        
+       
+    }
+
+
+}
+
+
