+++ quinzena6/LabenuForms/labenuforms/src/components/pagina4.jsx
@@ -1,0 +1,13 @@
+import React from "react"
+
+export default class Etapa4 extends React.Component {
+
+    render() {
+
+        return (
+            <div>
+                <h1>Processo Finalizado</h1>
+            </div>
+        )
+    }
+}
